Refactor to consolidate shared svgDraw components
