+++ packages/mermaid/src/diagrams/mindmap/svgDraw.js
@@ -69,6 +69,7 @@ const defaultBkg = function (elem, node, section) {
     .attr('x2', node.width)
     .attr('y2', node.height);
 };
+
 const rectBkg = function (elem, node) {
   elem
     .append('rect')
@@ -77,6 +78,7 @@ const rectBkg = function (elem, node) {
     .attr('height', node.height)
     .attr('width', node.width);
 };
+
 const cloudBkg = function (elem, node) {
   const w = node.width;
   const h = node.height;
@@ -107,6 +109,7 @@ const cloudBkg = function (elem, node) {
     H0 V0 Z`
     );
 };
+
 const bangBkg = function (elem, node) {
   const w = node.width;
   const h = node.height;
@@ -138,6 +141,7 @@ const bangBkg = function (elem, node) {
     H0 V0 Z`
     );
 };
+
 const circleBkg = function (elem, node) {
   elem
     .append('circle')
